+++ plugins/qtcoinrave/qtcoinviewer.h
@@ -310,7 +310,9 @@ protected:
 
     virtual void _UpdateEnvironment();
 
-    bool SetFiguresInCamera(ostream& sout, istream& sinput);
+    bool _SetFiguresInCamera(ostream& sout, istream& sinput);
+    bool _SetWatermark(ostream& sout, istream& sinput);
+    void _AddWatermarkToImage(unsigned char* image, int width, int height);
 
     // selection and deselection handling
     static void _SelectHandler(void *, class SoPath *); 
@@ -352,6 +354,8 @@ protected:
 
     boost::array<SoText2*,2> _messageNodes;
     SoTranslation* _messageShadowTranslation;
+
+    boost::multi_array<uint32_t,2> _vwatermarkimage;
 
     bool _altDown[2];
     bool _ctrlDown[2];